+++ apollo-converter/src/main/java/com/apollographql/android/converter/OperationJsonAdapter.java
@@ -1,4 +1,4 @@
-package com.apollographql.android.converter.pojo;
+package com.apollographql.android.converter;
 
 import com.apollographql.android.api.graphql.Operation;
 import com.squareup.moshi.JsonAdapter;
