+++ apollo-converter/src/test/java/com/apollographql/android/converter/ApolloConverterFactoryTest.java
@@ -1,12 +1,10 @@
-package com.apollographql.android.converter.pojo;
+package com.apollographql.android.converter;
 
 import com.apollographql.android.api.graphql.Operation;
 import com.apollographql.android.api.graphql.Response;
 import com.apollographql.android.api.graphql.ResponseFieldMapper;
 import com.apollographql.android.api.graphql.ResponseReader;
-import com.apollographql.android.converter.ApolloConverterFactory;
-import com.apollographql.android.converter.CustomTypeAdapter;
-import com.apollographql.android.converter.pojo.type.CustomType;
+import com.apollographql.android.converter.type.CustomType;
 import com.squareup.moshi.Types;
 
 import org.junit.Before;
