+++ apollo-converter/src/test/java/com/apollographql/android/converter/ProductsWithDate.java
@@ -1,11 +1,11 @@
-package com.apollographql.android.converter.pojo;
+package com.apollographql.android.converter;
 
 import com.apollographql.android.api.graphql.Field;
 import com.apollographql.android.api.graphql.Operation;
 import com.apollographql.android.api.graphql.Query;
 import com.apollographql.android.api.graphql.ResponseFieldMapper;
 import com.apollographql.android.api.graphql.ResponseReader;
-import com.apollographql.android.converter.pojo.type.CustomType;
+import com.apollographql.android.converter.type.CustomType;
 
 import java.io.IOException;
 import java.lang.Object;
